# Optimize selectors matching