--- a/lib/floki/selector.ex
+++ b/lib/floki/selector.ex
@@ -143,6 +143,11 @@ defmodule Floki.Selector do
 
   defp do_classes_matches?(nil, _classes), do: false
 
+  defp do_classes_matches?(class_attr_value, [class | _])
+       when bit_size(class_attr_value) < bit_size(class) do
+    false
+  end
+
   defp do_classes_matches?(class_attr_value, classes) do
     classes -- String.split(class_attr_value, [" ", "\t", "\n"]) == []
   end