--- a/lib/floki/selector.ex
+++ b/lib/floki/selector.ex
@@ -153,8 +153,14 @@ defmodule Floki.Selector do
     class == class_attr_value
   end
 
+  defp do_classes_matches?(class_attr_value, [class]) do
+    Enum.member?(String.split(class_attr_value, [" ", "\t", "\n"]), class)
+  end
+
   defp do_classes_matches?(class_attr_value, classes) do
-    classes -- String.split(class_attr_value, [" ", "\t", "\n"]) == []
+    min_size = Enum.reduce(classes, -1, fn item, acc -> acc + 1 + bit_size(item) end)
+    can_match? = bit_size(class_attr_value) >= min_size
+    can_match? && classes -- String.split(class_attr_value, [" ", "\t", "\n"]) == []
   end
 
   defp attributes_matches?(_node, []), do: true