--- a/lib/floki/selector.ex
+++ b/lib/floki/selector.ex
@@ -78,11 +78,26 @@ defmodule Floki.Selector do
   def match?(%Comment{}, _selector, _tree), do: false
 
   def match?(html_node, selector, tree) do
-    id_match?(html_node, selector.id) && namespace_match?(html_node, selector.namespace) &&
-      type_match?(html_node, selector.type) && classes_matches?(html_node, selector.classes) &&
+    can_match_combinator?(html_node, selector.combinator) &&
+      id_match?(html_node, selector.id) &&
+      namespace_match?(html_node, selector.namespace) &&
+      type_match?(html_node, selector.type) &&
+      classes_matches?(html_node, selector.classes) &&
       attributes_matches?(html_node, selector.attributes) &&
       pseudo_classes_match?(html_node, selector.pseudo_classes, tree)
   end
+
+  defp can_match_combinator?(_node, nil), do: true
+
+  defp can_match_combinator?(
+         %HTMLNode{children_nodes_ids: []},
+         %Selector.Combinator{match_type: match_type}
+       )
+       when match_type in [:child, :descendant] do
+    false
+  end
+
+  defp can_match_combinator?(_node, _combinator), do: true
 
   defp id_match?(_node, nil), do: true
   defp id_match?(%HTMLNode{attributes: []}, _), do: false