--- a/lib/floki/selector/attribute_selector.ex
+++ b/lib/floki/selector/attribute_selector.ex
@@ -63,7 +63,7 @@ defmodule Floki.Selector.AttributeSelector do
     s.attribute
     |> get_value(attributes)
     # Splits by whitespaces ("a  b c" -> ["a", "b", "c"])
-    |> String.split([" ", "\t", "\n"])
+    |> String.split([" ", "\t", "\n"], trim: true)
     |> Enum.any?(fn v -> String.downcase(v) == selector_value end)
   end
 
@@ -103,8 +103,8 @@ defmodule Floki.Selector.AttributeSelector do
 
   def match?(attributes, s = %AttributeSelector{match_type: :includes, value: value}) do
     get_value(s.attribute, attributes)
-    |> String.split([" ", "\t", "\n"])
-    |> Enum.any?(fn v -> v == value end)
+    |> String.split([" ", "\t", "\n"], trim: true)
+    |> Enum.member?(value)
   end
 
   def match?(attributes, s = %AttributeSelector{match_type: :dash_match}) do